Changed a integer value to a float
Former-commit-id: 4f3e703a360fdadc36f112bee1f57e5bc399b171

+++ ITIPD/extractor/models.py
@@ -45,5 +45,5 @@ class AccessLog(models.Model):
 class Agreement(models.Model):
     first = models.ForeignKey(MappingUnitToUser, related_name="first")
     second = models.ForeignKey(MappingUnitToUser, related_name="second")
-    percentage_by_types = models.IntegerField(default=0)
+    percentage_by_types = models.FloatField(default=0)
     percentage_by_chars = models.IntegerField(default=0)
